backend/requirements.txt
# Core dependencies
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-unstructured>=0.1.0
unstructured>=0.10.0
unstructured[local-inference]>=0.10.0

# PDF processing
pymupdf>=1.26.4
pypdf>=3.0.0
reportlab>=3.6.0

# OCR & Image processing
paddlepaddle>=2.5.0
paddleocr>=2.7.0
pytesseract>=0.3.10
pillow>=9.0.0
opencv-python>=4.8.0

# Data processing
html2text>=2025.4.15
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=1.5.0

# Development
jupyter>=1.0.0
notebook>=6.5.0
python-dotenv>=1.0.0

# System requirements (must be installed separately):
# - poppler (PDF rendering)
# - tesseract (OCR engine)